cleaned up our component
