+++ app/components/Layout.js
@@ -1,23 +1,14 @@
 import React from 'react';
 
-const data = [
-  { name: 'Zack Siri', email: 'zack@codemy.net' },
-  { name: 'Savika', email: 'savika@example.com' },
-  { name: 'Chad', email: 'chad@example.com' },
-];
-
-const Contact = props =>
-  <div className='pure-u-1-3'>
-    <h2>{props.name}</h2>
-    <p>{props.email}</p>
-  </div>;
+import Contact from './Contact';
+import data from './data';
 
 class Layout extends React.Component {
   render() {
     return (
       <div id='Layout' className='pure-g'>
         {data.map(info =>
-          <Contact {...info} />
+          <Contact key={info.id} {...info} />
         )}
       </div>
     );
