+++ app/components/data.js
@@ -1,0 +1,7 @@
+const data = [
+  { id: 1, name: 'Zack Siri', email: 'zack@codemy.net' },
+  { id: 2, name: 'Savika', email: 'savika@example.com' },
+  { id: 3, name: 'Chad', email: 'chad@example.com' },
+];
+
+export default data;
